Move all tests to tests/subdirectory
