+++ tests/components/binding.gyp
