+++ tests/integration/zk_test_watcher.js
@@ -1,4 +1,4 @@
-var { constants, ZooKeeper: ZK } = require("../lib/index");
+var { constants, ZooKeeper: ZK } = require("../../lib/index");
 var assert = require('assert');
 var util = require('util');
