+++ tests/integration/zk_test_watcher_session.js
@@ -1,4 +1,4 @@
-var { constants, ZooKeeper: ZK } = require ("../lib/index");
+var { constants, ZooKeeper: ZK } = require ("../../lib/index");
 
 if (process.argv.length < 2)
     throw new Error ("must supply number of  sessions (optionally)");
