+++ tests/integration/zk_worker.js
@@ -1,7 +1,7 @@
 var assert = require('assert');
 var util = require('util');
 
-var { constants, ZooKeeper: ZK } = require ('../lib/index');
+var { constants, ZooKeeper: ZK } = require ('../../lib/index');
 
 onmessage = function(msg) {
     postMessage({ echo : msg });
